pyproject.toml
[project]
name = "candypay-creditcard-ocr"
version = "0.1.0"
description = ""
authors = [
    {name = "CodePsy-2001",email = "codepsy2001@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (==0.116.1)",
    "uvicorn (==0.35.0)",
    "pydantic-core (==2.35.2)",
    "pillow (>=11.3.0,<12.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "black (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
